requirements.txt
tensorflow-gpu
lxml
nltk
numpy
tqdm
dotmap
tensorflow_hub
pydot
graphviz
